refactor(Likert): split design, experiment_logic
